+++ src/main/java/com/xxxmkxxx/customgui/client/hierarchy/node/animations/Animation.java
@@ -2,5 +2,5 @@ package com.xxxmkxxx.customgui.client.hierarchy.node.animations;
 
 @SuppressWarnings("unused")
 public interface Animation {
-    void addFrame(int timeUnit, AnimationFrame frame);
+    void addFrame(long timeUnit, AnimationFrame frame);
 }
